Refacter application main flame
Devide drawer and toolbar
